Fix handling for timestamps > 4 bytes.
The Python implementation serialises the timestamp as an 8-byte integer
(using pack with '>Q'[1]), and then trims leading zeroes[2].

This implementation was assuming a fixed 4-byte representation, which
only works because all current timestamps are 4 bytes once leading
zeroes are removed. This updates it to match the Python implementation,
which means it won't break when the current timestamp overflows 4 bytes.

[1]https://github.com/pallets/itsdangerous/blob/2.2.0/src/itsdangerous/encoding.py#L44
[2]https://github.com/pallets/itsdangerous/blob/2.2.0/src/itsdangerous/encoding.py#L50

+++ itsdangerous.go
@@ -44,6 +44,6 @@ var NowFunc = time.Now
 
 // Returns the current timestamp.  This implementation returns the
 // seconds since January 1, 1970 UTC.
-func getTimestamp() uint32 {
-	return uint32(NowFunc().Unix())
+func getTimestamp() int64 {
+	return NowFunc().Unix()
 }
